feat: Add cobe library for creating a globe widget
This commit adds the "cobe" library as a new dependency in the package.json file. The library is used to create a globe widget in the "GlobeWidget" component. The component sets up a canvas element and initializes the globe with various parameters such as size, colors, and markers. The globe is rendered using the "createGlobe" function from the "cobe" library. When the component is unmounted, the globe is destroyed to clean up resources.